EasyOCR/requirements.txt
easyocr
pillow
pymupdf
